+++ acc_app_optimisation/job_control/rl/execute.py
@@ -3,7 +3,7 @@ from logging import getLogger
 from pathlib import Path
 
 import gym
-from cernml.coi.unstable import cancellation
+from cernml.coi import cancellation
 from gym.envs.registration import EnvSpec
 
 from ...envs import make_env_by_name